Baselines/pr 5168 base (#1593)
* reviewed

* aliases

* reviewed

* reviewed

* Baselines for https://github.com/linq2db/linq2db/pull/5168 (#1590)

* changes

* [Linux / ClickHouse MySql] baselines

* [Linux / ClickHouse Driver] baselines

* [Linux / ClickHouse Octonica] baselines

---------


---------

+++ Firebird.2.5/Tests/Linq/ConditionalTests/Tests.Linq.ConditionalTests.Conditional_DoNotEvalueBothBranches(Firebird.2.5).sql
@@ -1,6 +1,6 @@
 ﻿-- Firebird.2.5 Firebird
 
 SELECT
-	CAST(20 AS Float)
+	CAST(20 AS DOUBLE PRECISION)
 FROM rdb$database
