+++ Firebird.2.5/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ToSqlFloat(Firebird.2.5).sql
@@ -1,7 +1,7 @@
 ﻿-- Firebird.2.5 Firebird
 
 SELECT
-	CAST("t"."MoneyValue" AS Float)
+	CAST("t"."MoneyValue" AS DOUBLE PRECISION)
 FROM
 	"LinqDataTypes" "t"
